Remove SpatialPose class
After discussion, we (Sherm and I) concluded that pose was better expressed
simply as an Isometry3. And rather than changing SpatialPose underlying
representation, it would be more convenient to simply replace
SpatialPose with what it is. The syntactic sugar of extracting position
and orientation quaternion from the object didn't seem worthwhile.