Revamping the proximity tests
Creating the "tangent" box test. It's a test that conceptually collides a
posed box against a "tangent plane". That "tangent" plane can be replaced
with arbitrary convex geometries that are tangent to the plane *at* the
point of contact. They should all produce the same collision result. This
tests that.