Expose collision model updates through RBT
In the RigidBodyTree regime, collision queries pass through the
RigidBodyTree. Traditionally, prior to queries, the RBT collision elemets
are updated from a KinematicsCache instance (computed by the RBT).

However, for applications that want to use RBT strictly for its collision
detection model, there is an alternate workflow. Callers can individually
invoke `updateCollisionElements` on a body with an arbitrary world pose
provided (as opposed to computing the KinematicsCache). This propagates the
pose to the collision element, but not into the collision model. This method
maps the poses in the collision into the collision model, completing the
alternate workflow.

TODO: Add unit test of the proposed workflow