Simplify sphere_vs_box
Align the box's frame with the world frame (maintaining the relationship
between the box and sphere). This makes the answer to reason about and
reproduces the error (the normal is still in the -x direction instead of
the +z direction).

(Incidentally, expand the information that is printed out when reporting
contacts to lcm.)