Add rendering camera to the bowling ball scenario
This adds the rgbd camera to the bowling ball scenario. It includes the
ability to save rendered images to disk (and a system that does just that).
The renderer used is the vtk renderer as a baseline case.

This includes a *fix* on the rgd_camera and a tweak of the ospray.