Reworking geometry/frame indices
The old indices were overworked -- the engine index *was* the index of the
object in the sets of anchored and geometry data. Those are now two
different quantities.

So, I've created a new "InternalIndex" which always is the index of the
Internal(Frame|Geometry|AnchoredGeometry) in its relevant datasets. This
becomes the value that gets provided to the proximity engine to map back to
the colliding objects.