Introduce MeshPainterSystem
The system uses a painter object to paint on a texture
associated with a canvas object. This system owns the texture
image and provides it on a single output port. In the current
incarantion:

1. The painter object is represented as a VolumeMesh and
   only Cylinder and Box are currently supported.
   The box illustrates the impact of having as sparse a volume
   mesh as possible -- only 6 tets. Over a cylinder with a
   reasonable amout of surface resolution, this is 2-3X
   faster.
2. The canvas must be a Mesh -- and the OBJ file must contain
   texture coordinates.

The painter system's geometric relationships are coupled with
geometries in the SceneGraph via their GeometryIds. The system
queries SceneGraph for their current poses and then collides
the two into a contact patch -- the patch is painted into the
texture image via its texture coordinates.

Known issues:
  1. Triangles whose texture coordinate extend over texture
     boundaries draw incorrectly. (Rather than being drawn
     with respect to periodic boundaries, they draw across
     the domain of the image.)
  2. This had to largely copy and paste code from
     mesh_intersection.h wholesale; it needed a different kind
     of sampling on the contact surface. The original code
     lacked the flexibility required.