Reformulating benchmark
TODO: With a single call, load the scene, run through a bunch of image
 sizes, and dump out a full table (something amenable to easy analysis --
 making charts, etc.)

Prepare a slide or two with those charts, highlighting the start up costs
of VTK (and to a lesser extent GL). Revisit FL with their analysis of cost.
Touch base with Huihua to see what he knows on the subject.