Remove TriggerTest from leaf_system_test.cc
This is part of an effort to remove antiquated event declarations (i.e.,
overriding various system Do* methods related methods instead of simply
declaring events.

In the case of this unit test, it has grown so stale as to not provide
value. It overrides DoPublish() and DoGetPerStepEvents(). These are two
*abstract* methods that LeafSystem itself overrides. So, we've removed
those two parts from being tested.

Given invocations of Publish, the *only* remaining method implemented in
LeafSystem that is *actually* being tested is DispatchPublishHandler().
That method is implicitly tested in all other tested invocations of
Publish(), so there's very little point in testing it here.

Current changes reduce it to a simpler version of its current incarnation:

Two choices:

1. Just kill the test (I've pinged Sherm to get his opinion)
2. Extend it to test the dispatch of simultaneous events across all event
   types). This would require likewise figuring out the significance of
   calling the Force Publish? And the significance of confirming that
   force publish events aren't included in the event collection.