"Clean up" soft tet meshes
Examines a tet mesh and "cleans up" interior tet faces whose vertices all
lie on the boundary. This is a "bad" thing because it leads to tets with
bad field properties.

This required making SortedTriplet a valid map key. And, while I was there,
I added structured binding support, because, why not?

TODO:
 - SortedTriplet needs test on the new behavior