Using the bounding box approach
This doesn't do the greatest framing always. The object will always be 100%
in frame, but not necessarily well framed.

This is because I'm ignoring *perspective*. The bounding box could be
inflated vertically because some distance point is high. But because of
perspective, that distance point converges to the center point (vanishing
point) and that leaves a void in the image space.