Explore inputs to resolve issue 18704
This creates a simple program that iterates through all of the objs posted
in the issue. It uses the poses and sphere measurements to attempt to
recreate the crash.

A report is output of all results.

This branch is referenced by the issue, so I'll keep it alive. If I chose
to delete it, update the issue indicating the branch is no more.