Add mtl to kuka meshes
The kuka arm wasn't appearing. By creating an .mtl file I was able to get
it display. Now with a host of warnings because we've applied colors in
the urdf file.

The materials shouldn't be necessary.

What I need to do is make some backwards compatible changes. It should
still work if all we have is rgba.

Two parts:

1. I shouldn't spew warnings yet. Those will come when we decide to flip the
   switch and no longer need drake materials.
2. Without an mtl file, things should render.

TODO:
 - Carefully rebase and finish off