Patch VTK glTF parsing
This addresses two different problems related to VTK making erroneous
assumptions that if *any* texture was present, then the base color texture
must be present. For example, omitting a base color texture but including
a roughness/metallic texture would do the following:

1. VTK would print warnings about multiple sets of UV coords (in reality
there was only a single set). Simply initializing the TextureInfo data
fully alleviated this (note: this is not a bullet-proof solution because an
adversarial glTF could have a single set of UVs, but define them in the
set UV2 (instead of UV0).
2. Texture maps were not applied if there was no base color texture. Now,
the declared textures are handled independently.