Create shadow test bed
This takes the Shadows.cxx example and moves it into Drake. (Guaranteeing
it's using the same VTK as Drake is). The example's interactor doesn't work
properly -- not sure why.

This also adds an equivalent program that explicitly uses Drake constructs.
The intent is that it should produce identical images to the non-Drake
version. The two can be used in conjunction to "easily" assess shadow
casting in Drake.