Changing transforms
Previously, OpenGlGeometry was expressed in the same frame as the drake
geometry (with the possible exception of scaling it up). This is
incompatible with glTF files. A node can be arbitrarily posed within the
file.

Instances now calculate their own node-to-world transforms for both
position and normals and those are passed directly to the shaders
(simplifying the shader code).

This supplanted the Part::T_GN field. When removing Part::T_GN Part no
longer had any value and it was simply removed.
TODO: Part now has an used T_GN property; in fact, Part is now just an
instance. Change that.