A case where mass is good but inertia is bad
We compute mesh spatial inertia via divergence theorem. But to get a proper
inertia tensor, the mesh must be a properly closed manifold. We know it's
possible to create a spatial inertia with negative mass. The question is
could one be created with positive mass but negative inertia. This test
does exactly that.

However, it's not a legitimate unit test, becuase the test can't actually
distinguish the reason for validating not throwing. If we *really* wanted
to test it, we'd have to get back the *invalid* spatial inertia and poke
at its values directly (without throwing). It doesn't seem like it's worth
changing a *public* API to include a "don't validate" flag just for testing.

TODO:
 - Paul also thought about a scenario where the mass and inertia are both
   physically valid, but the CoM doesn't lie in the mesh's convex hull.
   - To test this, we'd want:
     - A version of MakeConvexHull that takes a TriangleSurfaceMesh
     - the ability to test if a point is inside or outside a convex hull.