Yaml serialization includes std::filesystem::path and vector<uint8_t>
Path member can now be serialized like other basic member types.

Serializing something of type vector<uint8_t> is interpreted as a byte
string. If all bytes are printable, it is equivalent to a string. If there
are any non-printable characters, it is converted to base64 and prefixed
with YAML's !!binary tag.

MemoryFile makes use of this to serialize its file contents (it creates
a temporary vector<uint8_t> for serialization to trigger the potential
binary yaml encoding.

TODO:
 0. Failing tests in //common/yaml:json_test, WriteScalars
 1. I need unit tests in python.
 2. Doing: `content: !binary lskdjflskdjf` is different in python vs C++.
    - python complains that there is no constructor for !binary.
    - C++ ignores !binary.
    Resolve this discrepancy when operating on undefined local tags.
 3. Prep PR that has nothing to do with rendering.