Modify how timers coordinate between the two engines
I had previously put a call to clearAllTimers() in the render engines.
This is generally a *bad* thing. I did it so that I wouldn't get reports
with data from the wrong engine.

The simpler thing to do is to rely on reset, and omit any row that has
no laps (no laps --> no data).