Add flag to test validity of experimental methodology
A concern was raised that simply using the same sphere over and over would
lead to results that didn't necessarily inform us of what to expect in
real-world scenarios.

My position was that it did. Particularly as VTK currently doesn't do any
instancing. Each geometry is its own independent bag of triangles and
textures. Whlle it's *possible* that the GPU could do some things to
recognize that it's getting textures and vertex buffers with the same
data, it is unlikely that it is occurring.

Running with the flag seems to suggest that my hypothesis is correct.

The only thing missing, is large scale. It's only 40 objects and my
experiments went all the way up to 1000 objects (well, 960). It may weel
be that differences would become more apparent with larger scale.

I'm leaving this as an exercise for the future.