Counters for candidates and observed
Note: this uses a different timing methodology. The "CalcContactPolygon"
and "AddPolygon" methods get called in parallel. The timer has tno
affordances for being called in parallel.

Instead, I use the lap timers to simply count the number of candidates I
had and the number of actual elements created. They should map to the old
numbers.

The reported *TIMES* are not to be trusted.